Simplify install instructions in README
This commit removes the installation instructions for installing the
repository from a local copy which would lead to errors as
setuptools_scm needs the repository .git folder to function. Therefore,
it is assumed that the user has Git installed on their system. However,
if they don't then a link to the website is provided.