Remove get_sample_points from Airfoil
This commit removes the get_sample_points method from the Airfoil class
as well as its respective test. This is because getting normalized
sample points is not a responsibility of an Airfoil, it is instead
dependent on a specific panel method. Therefore, this responsibility is
left for the Solver to implement.