Add ability to get ith panel from a Panel2D array
This commit modifies the __getitem__ magic method to treat integer
indices as a desire to retrieve the ith panel from a Panel2D array.
Furthermore, for all other indices it returns a view of the data as a
Point2D array which makes it more explicit that a Panel2D object
contains a set of n_panels + 1 nodes.

Summary of Changes:
- Overrode __getitem__ magic method in Panel2D to return the ith panel
  when an integer index is used
- Added additional tests to cover out of bound errors and ensure support
  for iteration