Add ability to load an airfoil from a file
This commit adds the ability to create an Airfoil instance from airfoil
data files. The difficulty of creating an airfoil definition from a
list of discrete points is that for the correct function of the
panel method, airfoil surface points need to be queried at
arbitrary locations.

To handle this requirement the upper and lower surfaces of the airfoil
are splined separately. Furthermore, a mean camber line as well as
camber detection algorithm are added to comply with the requirements of
the Airfoil Abstract Base Class (ABC).

Summary of Changes:
- Added FileAirfoil ABC along with the UIUCAirfoil specialization to
  define Airfoil instances from airfoil coordinate files. This respects
  the Open-Closed Principle and file formats can be added easily using
  the FileAirfoil ABC.
- Moved NACA4Series.plot method into the Airfoil ABC and created a
  specialization within the NACA4Series class to handle the generation
  of a plot title
- Added UIUCAirfoil to be directly importable from the geometry package
- Added unit tests to check basic function of the UIUCAirfoil class