Implement ThinFlowSolution and lift coefficient
This commit implements a ThinFlowSolution to handle the differences of
post-processing the lumped vortex panel method. Furthermore, it adds
a lift-coefficient calculation for both the new ThinFlowSolution as
well as the ThickFlowSolution.

Additionally, the ability to plot either a single AoA, a subset of AoAs,
or the entire range of AoAs is implemented with use of the
enumerate_alpha method.

Finally, the property names of the ThickFlowSolution are improved to
better represent the physical meaning of the normalized induced
velocities used to determine the pressure and lift coefficients.

Summary of Changes:
- Implement ThinFlowSoltuion for the lumped vortex panel method
- Add ability to plot a single AoA, a subset of the provided AoAs,
  or the entire range of AoAs
- Implement lift coefficient calculation for both the ThinFlowSoltuion
  and the ThickFlowSolution
- Improve method names of the ThickFlowSolution and complete docstrings
- Correct docstring of the FlowSolution base class to remove mention
  of old density and velocity arguments
- Make abstract methods in the FlowSolution class to require
  implementation of delta Cp, Cp, and Cl