Add parsed NACA0015 data from assignment PDF
This commit Closes #12 by adding the table data from the assignment
PDF as a multi-line text file. This is then parsed into a dictionary
by means of splitting the long list of numbers into chunks which
represent a single row in the assignment PDF. To demonstrate how the
data is contained in the dictionary a temporary plot is also added.

Summary of Changes:
- Added NACA0015 data from assignment PDf as a .txt file
- Parsed the .txt file into a dictionary
- Added a temporary plot to show usage of the dictionary