Rename package from GammaPy to NumFoil
This commit Closes #9 by renaming GammaPy to NumFoil which places
emphasis on the usage of Numpy and Numba to vectorize the calculations.
It also has a nice property that it can be interpreted as "Numerical"
Foil which has more significance to more people compared to "Gamma"
which was an arbitrary name.

Summary of Changes:
- Renamed main package to NumFoil from GammaPy
- Propogated changes across assignment tasks, tests, and the report
- Fixed old repository URLs in README.md